requirements.txt
streamlit>=1.31.0
supabase>=2.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
